requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2-binary==2.7.6
PyYAML==3.13
requests==2.20.1
six==1.11.0
snap-micro==0.9.91
SQLAlchemy==1.2.13
SQLAlchemy-Utils==0.33.6
urllib3==1.24.1
Werkzeug==0.14.1
